video/console: Add dmi quirk table for x86 systems which need fbcon rotation
Some x86 clamshell design devices use portrait tablet screens and a
display engine which cannot rotate in hardware, so we need to rotate
the fbcon to compensate.

This commit adds a DMI based quirk table which is initially populated with
4 such devices: The Asus T100HA, GPD Pocket, the GPD win and the I.T.Works
TW891, so that the console comes up in the right orientation on these
devices OOTB.

Unfortunately these (cheap) devices also typically have quite generic DMI
data, so we match on a combination of DMI data, screen resolution and a
list of known BIOS dates to avoid false positives.

Suggested-by: Jean Delvare <jdelvare@suse.de>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Jean Delvare <jdelvare@suse.de>
---
Changes in v2:
-Extend the commit message explaining why dmi_check_system does not work
 and this commit exports and uses dmi_matches instead

Changes in v3:
-The exporting of dmi_matches has been split out in a separate patch
-Use u32 for width/height in struct fbcon_dmi_rotate_data to match the
 type of fb_var_screeninfo.xres / .yres

Changes in v4:
-Use dmi_first_match instead of exporting dmi_matches
-Use DMI_EXACT_MATCH
-Add the Asus T100HA to the list of devices needing rotation

Changes in v5:
-Add Reviewed-by: Jean Delvare <jdelvare@suse.de>
-Add extra BIOS-date to known GDP pocket BIOS dates