staging: typec: fusb302: Add support for fcs,vbus-regulator-name device-property
On devicetree platforms the fusb302 dt-node will have a vbus regulator
propery with a phandle to the regulator.

On ACPI platforms, there are no phandles and we need to get the vbus by a
system wide unique name. Add support for a new "fcs,vbus-regulator-name"
device-property which ACPI platform code can set to pass the name.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>